Rename module, update CLI commands; Fix import config.py in current folder
